House keeping, and organization
House keeping;
Renamed __del__ to close, and have it called from write() since file stream doesn't always clear buffers until close. (also if the application closes before garbage collection, __del__ might never be called)
